app/requirements.txt
flask
pytest
pytest-cov
codecov
facebook-sdk==3.1.0